\name{fPortfolio-package}


\alias{fPortfolio}


\docType{package}


\title{Portfolio Design, Optimization and Backtesting}


\description{

    The Rmetrics "fPortfolio" package is a very powerful collection
    of functions to optimize portfolios and to analyze them from
    different points of view.

}


\details{

    \tabular{ll}{
        Package:    \tab fPortfolio\cr
        Type:       \tab Package\cr
        Date:       \tab 2011\cr
        License:    \tab GPL Version 2 or later\cr
        Copyright:  \tab (c) 1999-2011 Diethelm Wuertz and Rmetrics Association\cr
        URL:        \tab \url{http://www.rmetrics.org}
    }
    
}


\references{

Wuertz, D., Chalabi, Y., Chen W., Ellis A. (2009);
    \emph{Portfolio Optimization with R/Rmetrics}, 
    Rmetrics eBook, Rmetrics Association and Finance Online, Zurich.
    
}


\keyword{models}

