\name{Portfolio}

\alias{portfolio}

\title{Portfolio Modelling, Optimization and Benchmarking}


\description{
  
    A collection and description of functions for
    portfolio modelling, pptimization and benchmarking.
    \cr
    
}
    

\details{

    \bold{Assets Modelling:}
    \cr\cr
    Functions for the analysis, simulation and parameter estimation
    of financial assets.
    \cr
       
    \bold{Portfolio Optimization:}
    Functions for the optimization of portfolios including the mean-variance
    Markowitz approach, "MV", the lower partial moment portfolio approach 
    "PM", and the conditional value-at-risk approach "CVaR".
    \cr\cr
    
    \bold{Benchmarkung:}
    Functions for the scenario analysis, portfolio performance measures,
    and benchmarking.
    \cr\cr
    

}


\keyword{models}
