\name{PortfolioConstraints}


\alias{PortfolioConstraints}

\alias{setConstraints}
\alias{getConstraints} 
\alias{getConstraints.constraintsMatrix}        


\title{Portfolio Constraints}


\description{
  
    A collection and description of ...
    \cr
    
    The functions are:
    
    \tabular{ll}{
    \code{setConstraints} \tab Transforms constraint strings into a list value, \cr
    \code{getConstraints} \tab Transforms a constraint list value into strings. }
    
}
    

\usage{
setConstraints(data, spec = portfolioSpec(), constraintsStrings = NULL) 
getConstraints(object)

}


\details{

    \bold{How to define constraints:}
    \cr\cr
    Describe the ...
    \cr

}


\arguments{

    \item{constraintsStrings}{
        ...
        }
    \item{data}{
        ...
        }
    \item{object}{
        ...
        }
    \item{spec}{
        ...
        }

}


\value{
  
    \code{setConstraints} 
    \cr
    returns a ...
    \cr

    \code{getConstraints} 
    \cr
    returns a ...
  
}


%\references{}


\seealso{

    \code{...}.

}


\author{

    Diethelm Wuertz and Oliver Greshake for the Rmetrics port.
    
}


\examples{

## setConstraints -
   # ...
   
## getConstraints -
   # ...
 
}


\keyword{models}

