\name{PortfolioPlots}

\alias{PortfolioPlots}

\alias{frontierPlot} 
    
\alias{weightsPlot}      
\alias{weightsPie}       
\alias{attributesPlot}   
  
\alias{covEllipsesPlot}    


\title{Portfolio Plots}


\description{
  
    A collection and description of ...
    \cr
    
    The functions are:
    
    \tabular{ll}{
    \code{frontierPlot} \tab Plots efficient frontier, \cr              
    \code{weightsPlot} \tab Plots staggered weights, \cr
    \code{weightsPie} \tab Plots staggered weights, \cr
    \code{attributesPlot} \tab Plots weighted means, \cr
    \code{covEllipsesPlot} \tab Plots covariance ellipses. }
    
}
    

\usage{
frontierPlot(object, frontier = c("both", "lower", "upper"),
    col = c("black", "grey"), add = FALSE, ...) 
    
weightsPlot(object, col = NULL, legend = FALSE)      
weightsPie(object, col = NULL, box = TRUE, \dots)       
attributesPlot(object, col = NULL, legend = FALSE)   
  
covEllipsesPlot(x = list(), \dots)    

}


\details{

    \bold{The frontier Plot:}
    \cr\cr
    Describe the data set ...
    \cr
    
    
    \bold{Weights and Attributes Plots:}
    \cr\cr
    Describe the data set ...
    \cr
    
    
    \bold{Covariance Ellipses Plot:}
    \cr\cr   
    Describe the data set ...

}


\arguments{

    \item{add}{
        ...
        }
    \item{box}{
        ...
        }
    \item{col}{
        ...
        }
    \item{frontier}{
        ...
        }
    \item{legend}{
        ...
        }
    \item{object}{
        ...
        }
    \item{x}{
        ...
        }
    \item{\dots}{
        ...
        }

}


\value{
 
    ...
  
}


%\references{}


\seealso{

    \code{xxx}.

}


\author{

    Diethelm Wuertz and Oliver Greshake for the Rmetrics port.
    
}


\examples{

## fun -
   # fun()
 
}


\keyword{models}

