\name{fPFOLIOVAL-class}


\alias{fPFOLIOVAL}
\alias{class-fPFOLIOVAL}

\alias{fPFOLIOVAL-class}

\alias{show,fPFOLIOVAL-method}


\title{Values of Portfolio Frontiers}


\description{
  
    Specifies portfolio Optimized Values.
    
}


\usage{       
\S4method{show}{fPFOLIOVAL}(object)
}


\arguments{

    \item{object}{
        an S4 object of class \code{fPFOLIOVAL}.
        }

}       



\references{

Wuertz, D., Chalabi, Y., Chen W., Ellis A. (2009);
    \emph{Portfolio Optimization with R/Rmetrics}, 
    Rmetrics eBook, Rmetrics Association and Finance Online, Zurich.
    
}


\examples{
## 

}


\keyword{models}

