\name{covEllipsesPlot}


\alias{covEllipsesPlot}


\title{Covariance Ellipses Plot}


\description{

    Plots covariance ellipses.

}


\usage{
covEllipsesPlot(x = list(), \dots)
}


\arguments{

    \item{x}{
        a list of at least two covariance matrices.
        }
    \item{\dots}{
        optional arguments to be passed.\cr
        }

}


\details{

    This plot visualizes the difference between two or more covariance matrices.
    It is meant to compare different methods of covariance estimation.

}



%\references{}


\author{

    Diethelm Wuertz and Oliver Greshake for the Rmetrics port.

}


\examples{
## ---
}


\keyword{models}

