\name{data-sets}


\alias{dataSets}

\alias{GCCINDEX.DF}
\alias{SPISECTOR.DF}
\alias{SWX.DF}
\alias{LPP2005.RET.DF}
\alias{SMALLCAP.RET.DF}

\alias{GCCINDEX}
\alias{SPISECTOR}
\alias{SWX}
\alias{LPP2005}
\alias{SMALLCAP}

\alias{GCCINDEX.RET}
\alias{SPISECTOR.RET}
\alias{SWX.RET}
\alias{LPP2005.RET}
\alias{SMALLCAP.RET}

\alias{ECON85}
\alias{ECON85LONG}


\title{Assets Data Sets}


\description{
  
    Example data sets for portfolio optimization.
    
}
    

\usage{
ECON85
ECON85LONG

GCCINDEX
SPISECTOR
SWX
LPP2005
SMALLCAP

GCCINDEX.RET
SPISECTOR.RET
SWX.RET
LPP2005.RET
SMALLCAP.RET
}


%\details{
%}


\value{
  
    an object of class \code{"timeSeries"}.
  
}


\references{

Wuertz, D., Chalabi, Y., Chen W., Ellis A. (2009);
    \emph{Portfolio Optimization with R/Rmetrics}, 
    Rmetrics eBook, Rmetrics Association and Finance Online, Zurich.
    
}


\keyword{datasets}

