\name{getVal}


\alias{getAlpha.fPFOLIOVAL}
\alias{getCovRiskBudgets.fPFOLIOVAL}
\alias{getNFrontierPoints.fPFOLIOVAL}
\alias{getPortfolio.fPFOLIOVAL}
\alias{getRiskFreeRate.fPFOLIOVAL}
\alias{getStatus.fPFOLIOVAL}
\alias{getTargetReturn.fPFOLIOVAL}
\alias{getTargetRisk.fPFOLIOVAL}
\alias{getWeights.fPFOLIOVAL}


\title{PortfolioVal Extractor Functions}


\description{
  
    Extracts information from an object of class fPFOLIOVAL. 
    
}
    

\usage{

\method{getAlpha}{fPFOLIOVAL}(object)
\method{getCovRiskBudgets}{fPFOLIOVAL}(object)
\method{getNFrontierPoints}{fPFOLIOVAL}(object)
\method{getPortfolio}{fPFOLIOVAL}(object)
\method{getRiskFreeRate}{fPFOLIOVAL}(object)
\method{getStatus}{fPFOLIOVAL}(object)
\method{getTargetReturn}{fPFOLIOVAL}(object)
\method{getTargetRisk}{fPFOLIOVAL}(object)
\method{getWeights}{fPFOLIOVAL}(object)

}


\arguments{

    \item{object}{
        an object of class \code{fPFOLIODATA}.
        }

}


% \details{

% }


\references{

Wuertz, D., Chalabi, Y., Chen W., Ellis A. (2009);
    \emph{Portfolio Optimization with R/Rmetrics}, 
    Rmetrics eBook, Rmetrics Association and Finance Online, Zurich.
    
}


\examples{
## 

}


\keyword{models}

