\name{portfolioTest}


\alias{portfolioTest}


\title{Portfolio Demo Examples}


\description{
  
    Runs portfolio demos as test cases.
    
}
    

\usage{  
portfolioTest(type = "MV", optimize = "minRisk", solver = "solveRquadprog", 
    constraints = c("LongOnly", "Short", "BoxGroup", "CovBudget"), 
    NFrontierPoints = 25)
}


\arguments{


    \item{type}{
        ...
        }    
    \item{optimize}{
        ...
        }
    \item{solver}{
        ...
        }
    \item{constraints}{
        ...
        }
    \item{NFrontierPoints}{
        ...
        }

}


%\value{}
  

%\references{}


\examples{
## ...
}


\keyword{models}

