\name{risk-surfaceRisk}


\alias{markowitzHull}
\alias{feasibleGrid}


\title{Surface Risk Analytics}


\description{
  
    Functions for surface risk analytics.
    
}


\usage{
markowitzHull(data, nFrontierPoints=50)
feasibleGrid(hull, trace=FALSE)
}


\arguments{

    \item{data}{data}
    \item{hull}{hull}
    \item{nFrontierPoints}{nFrontierPoints}
    \item{trace}{trace}
}


\references{

Wuertz, D., Chalabi, Y., Chen W., Ellis A. (2009);
    \emph{Portfolio Optimization with R/Rmetrics}, 
    Rmetrics eBook, Rmetrics Association and Finance Online, Zurich.
    
}


\keyword{models}


