\name{Rquadprog}


\alias{Rquadprog}
\alias{rquadprog}
\alias{quadprog}


\title{Quadratic Programming Solver}


\description{

    Solve Quadratic Programming Problem using 'quadprog' solver.
    
}


\usage{
rquadprog(Dmat, dvec, Amat, bvec, meq)
}


\arguments{
    
    \item{Dmat}{
        ...
        }
    \item{dvec}{
        ...
        }
    \item{Amat}{
        ...
        }
    \item{bvec}{
        ...
        }
    \item{meq}{
        ...
        }
        
}


%\details{}


%\note{}


\value{
    
    \item{\dots}{
        ...
        }
        
}


%\references{}


%\author{}


\keyword{optimize}

