\name{solve-environment}
 

\alias{Data}

\alias{portfolioObjective}
\alias{portfolioReturn}
\alias{portfolioRisk}


\title{Nonlinear Objective Presettings}


\description{
  
    Prests variables for Data, portfolioObjective, portfolioReturn,
    and portfolioRisk in the case of NL math programming of 
    portfolios.
    
}
    

\usage{
Data

portfolioObjective(weights)
portfolioReturn(weights)
portfolioRisk(weights)
}


\arguments{
  
    \item{weights}{
        a vector of portfolio weights 
        }
    
}


\references{

Wuertz, D., Chalabi, Y., Chen W., Ellis A. (2009);
    \emph{Portfolio Optimization with R/Rmetrics}, 
    Rmetrics eBook, Rmetrics Association and Finance Online, Zurich.
    
}


\keyword{optim}

