\name{utils-methods}


\alias{print.solver}


\title{Print Method for Solvers}


\description{

    S3 print method for mathematical programming solvers.

}


\usage{
\method{print}{solver}(x, \dots)
}


\arguments{

    \item{x}{x}
    \item{\dots}{optional arguments}

}


\references{

Wuertz, D., Chalabi, Y., Chen W., Ellis A. (2009);
    \emph{Portfolio Optimization with R/Rmetrics}, 
    Rmetrics eBook, Rmetrics Association and Finance Online, Zurich.
    
}


\keyword{models}

