\name{fPortfolio-package}


\alias{fPortfolio}


\docType{package}


\title{Portfolio Modelling, Optimization and Backtesting}


\description{

    The Rmetrics "fPortfolio" package is a very powerful collection
    of functions to optimize portfolios and to analyze them from
    different points ov view.
    
    The implemented portfolio models include the traditional 
    mean--variance Markowitz portfolio, robust variants of
    the Markowitz portfoio, and the mean-CVaR conditional 
    value-at-Risk portfolio.
    
    Optimization is possible by minimizing the risk if the return
    is specified.
    
    Linear box/group constraints can be specified.
    
    Depending on the model of the portfolio and the constraints
    a QP (quadratic programming) and a LP (linear programming) 
    solver are provided for optimization
    
    Several kinds of charts can be produced using graphics 
    tools to visualize the results. 

}


\details{

    \tabular{ll}{
        Package:    \tab fPortfolio\cr
        Type:       \tab Package\cr
        Date:       \tab 2009\cr
        License:    \tab GPL Version 2 or later\cr
        Copyright:  \tab (c) 1999-2008 Diethelm Wuertz and Rmetrics Association\cr
        URL:        \tab \url{http://www.rmetrics.org}
    }
    
}


\references{

Wuertz, D., Chalabi, Y., Chen W., Ellis A. (2009);
    \emph{Portfolio Optimization with R/Rmetrics}, 
    Rmetrics eBook, Rmetrics Association and Finance Online, Zurich.
    
}


\keyword{models}

