\name{PortfolioData}

\alias{PortfolioData}

\alias{dutchPortfolioData}
\alias{usPortfolioData}
\alias{sm132PortfolioData}


\title{Example Portfolio Data Sets}


\description{
  
    A collection and description of xxx
    \cr
    
    The data sets are:
    
    \tabular{ll}{
    \code{dutchPortfolioData} \tab xxx, \cr
    \code{usPortfolioData} \tab xxx, \cr
    \code{sm132PortfolioData} \tab xxx. }
    
}
    

\usage{
dutchPortfolioData() 
usPortfolioData()
sm132PortfolioData()
}


\details{

    \bold{Dutch Portfolio Data Set:}
    \cr\cr
    Describe the data set xxx
    \cr
    
    
    \bold{US Portfolio Data Set:}
    \cr\cr
    Describe the data set xxx
    \cr
    
    
    \bold{Scherer-Martin Data Set:}
    \cr\cr   
    Describe the data set xxx

}


\value{
  
    \code{dutchPortfolioData} 
    \cr
    returns a xxx
    \cr

    \code{dutchPortfolioData} 
    \cr
    returns a xxx
    \cr
    
    \code{dutchPortfolioData} 
    \cr
    returns a xxx 
  
}


\references{

N. N. (1985);
    \emph{Title},
    Scandinavian Journal of Statistics 12, 171--178. 

}


\seealso{

    \code{xxx}.

}


\author{

    Diethelm Wuertz and Oliver Greshake for the Rmetrics port.
    
}


\examples{

## usPortfolioData -
   # x = usPortfolioData()
 
}


\keyword{models}

