\name{PortfolioConstraints}


\alias{PortfolioConstraints}

\alias{setConstraints}
\alias{getConstraints} 
\alias{getConstraints.constraintsMatrix}        


\title{Portfolio Constraints}


\description{
  
    A collection and description of xxx
    \cr
    
    The functions are:
    
    \tabular{ll}{
    \code{setConstraints} \tab Transforms constraint strings into a list value, \cr
    \code{getConstraints} \tab Transforms a constraint list value into strings. }
    
}
    

\usage{
setConstraints(data, spec = portfolioSpec(), constraintsStrings = NULL) 
getConstraints(object)

}


\details{

    \bold{How to define constraints:}
    \cr\cr
    Describe the xxx.
    \cr

}


\arguments{

    \item{constraintsStrings}{
        xxx
        }
    \item{data}{
        xxx
        }
    \item{object}{
        xxx
        }
    \item{spec}{
        xxx
        }

}


\value{
  
    \code{setConstraints} 
    \cr
    returns a xxx
    \cr

    \code{getConstraints} 
    \cr
    returns a xxx
  
}


%\references{}


\seealso{

    \code{xxx}.

}


\author{

    Diethelm Wuertz and Oliver Greshake for the Rmetrics port.
    
}


\examples{

## setConstraints -
   # xxx
   
## getConstraints -
   # xxx
 
}


\keyword{models}

