\name{fFOLIOCON-class}


\alias{fPFOLIOCON}
\alias{fPFOLIOCON-class}


\title{Portfolio Constraints Handling}


\description{
  
    Creates a fPFOLIOCON object.
      
}


%\references{}


\author{

    Diethelm Wuertz and Oliver Greshake for the Rmetrics port.
    
}


\examples{
## ...
}


\keyword{models}

