\name{show-methods}


\alias{show-methods}
\alias{show,fPORTFOLIO-method}


\title{Portfolio Print Methods}


\description{

    show-methods.

}

\usage{
\S4method{show}{fPORTFOLIO}(object)
}


\arguments{

    \item{object}{
        an S4 object of class \code{fPORTFOLIO}.
        }

}


\references{

Wuertz, D., Chalabi, Y., Chen W., Ellis A. (2009);
    \emph{Portfolio Optimization with R/Rmetrics}, 
    Rmetrics eBook, Rmetrics Association and Finance Online, Zurich.
    
}


\keyword{models}

