\name{show-methods}


\alias{show-methods}
\alias{show,fPORTFOLIO-method}


\title{Portfolio Print Methods}


\description{

    show-methods.

}

\usage{
\S4method{show}{fPORTFOLIO}(object)
}


\arguments{

    \item{object}{
        an S4 object of class \code{fPORTFOLIO}.
        }

}


\keyword{models}

